+++ packages/terra-site/src/examples/time-input/TimeInputNoDefaultTime.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import TimeInput from 'terra-time-input';
+
+const timeInput = () => (<TimeInput name="time-input-default" />);
+
+export default timeInput;