+++ packages/terra-site/CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 Unreleased
 ----------
+- Moved DatePickers and TimeInput examples into their own files.
 
 1.0.0 - (June 28, 2017)
 ------------------